# gen/src/openapi_server/models/column_schema.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401
from gen.src.openapi_server.models.column_type import ColumnType


class ColumnSchema(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    ColumnSchema - a model defined in OpenAPI

        column_name: The column_name of this ColumnSchema [Optional].
    """

    column_name: Optional[str] = Field(alias="columnName", default=None)

    @validator("column_name")
    def column_name_min_length(cls, value):
        assert len(value) >= 1
        return value

ColumnSchema.update_forward_refs()