# gen/src/openapi_server/models/table_schema.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401


class TableSchema(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    TableSchema - a model defined in OpenAPI

        table_name: The table_name of this TableSchema.
    """

    table_name: str = Field(alias="tableName")

    @validator("table_name")
    def table_name_min_length(cls, value):
        assert len(value) >= 1
        return value

TableSchema.update_forward_refs()